Process aliases for names correctly
This package predefines a few aliases. If you want to get the g-code, you can now specify the virtual name '/toolpath' and it should point to that g-code.

This also normalizes the initial slash for the virtual paths. All virtual paths should now start with a slash.

Contributes to issue CURA-4872.