Use exact Python version when identifying the interpreter
The Python version that Cura is currently using is v3.5.7. However, all updated servers have now system Python versions installed by default, which are newer than that.
As a result, during CMake's configuration step the system Python is found first (since it's located in /usr/bin or /usr/local/bin) and, thus, the step results in error.
Using an exact version of Python and setting that variable during the configuration step is the most ideal solution for this issue.