CES-2235 Delay publication of the Charon service to the bus
Tweak the initialization order of the Charon DBus service. We should
initially create the `dbus.service.Object` *without* claiming a
well-known name on the bus. If we immediately claim a name within the
`FileService.__init__()` initializer, the DBus service will (briefly)
become visible on the DBus bus in a half-initialized state. The symptom
of this is clients seeing empty or incomplete introspection data on
`nl.ultimaker.charon`.

If, instead, we first allow `FileService` to completely finish its
`__init__()`, and then separately claim the well-known name in
`FileService.publish()`, this publishes the service to the bus
"atomically".